third_party/WebKit/LayoutTests/http/tests/devtools/elements/shadow/elements-panel-shadow-selection-on-refresh-2-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that elements panel preserves selected shadow DOM node on page refresh.


Running: setup

Running: testUserAgentShadowRoot
Page reloaded.
Selected node: '#shadow-root (user-agent)'

Running: testOpenShadowRootChild
Page reloaded.
Selected node: 'input'
